pyproject.toml
[tool.poetry]
name = "datamart-core"
version = "2.8.0a1"
description = ""
authors = ["sax <s.apostolico@gmail.com>"]

[tool.poetry.dependencies]
admin-extra-urls = "^2.1"
bump2version = "*"
celery = "*"
circus = "*"
cryptography = "==2.7"
django = "*"
django-adminactions = "^1.8"
django-adminfilters = "^1.1"
django-celery-beat = "*"
django-celery-email = "*"
django-concurrency = "*"
django-constance = "*"
django-cors-headers = "*"
django-countries = "*"
django-crashlog = "*"
django-crispy-forms = "*"
django-db-logging = "*"
django-dbtemplates = "*"
django-deprecate-fields = "*"
django-environ = "*"
django-filter = "*"
django-impersonate = "*"
django-jsoneditor = "*"
django-migration-linter = "*"
django-model-utils = "*"
django-mptt = "*"
django-picklefield = "*"
django-post-office = "*"
django-redis = "*"
django-regex = "*"
django-rest-framework-social-oauth2 = "*"
django-sql-explorer = "*"
django-strategy-field = "*"
django-sysinfo = "==1.4"
djangorestframework-csv = "*"
djangorestframework-gis = "*"
djangorestframework-jwt = "*"
djangorestframework-xml = "*"
djangorestframework-yaml = "*"
djangorestframework="^3.10.2"
drf-dynamic-serializer = "^2.1"
drf-extensions = "*"
drf-renderer-xlsx = "*"
drf-yasg = {version = "*",extras = ["validation"]}
humanfriendly = "*"
humanize = "*"
markdown = "*"
packaging = "*"
psycopg2-binary = "==2.7.7"
pygments = "*"
pyparsing = "*"
python = "^3.7"
python-social-auth = "*"
pyyaml = "^4.2b1"
rapidpro-python = "*"
raven = "*"
redis = "*"
sentry-sdk = "*"
social-auth-app-django = "*"
sqlparse = "*"
uwsgi = "*"
xhtml2pdf = "*"
django-storages = "^1.8"

[tool.poetry.dev-dependencies]
check-manifest= "*"
django-extensions = "*"
django-webtest = "*"
drf-api-checker = "^0.8"
factory-boy = "*"
freezegun = "*"
isort = "*"
pre-commit = "*"
pytest = "<5"
pytest-coverage = "*"
pytest-django = "*"
pytest-echo = "*"
pytest-ignore-flaky = "*"
pytest-pythonpath = "*"
tox = "*"
vcrpy = "*"
#fancycompleter = {file = "https://github.com/theY4Kman/fancycompleter/archive/0.8.tar.gz"}
flake8 = "*"
ipdb = "*"
bump2version = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
